+++ .gitignore
@@ -19,7 +19,7 @@ dist/
 /media/
 /node_modules/
 /static/
-
+/digital_contents/
 # Environments
 .env
 .venv
