+++ saleor/tests/e2e/checkout/test_checkout_create_from_order.py
@@ -1,119 +1,11 @@
 import pytest
 
-from ..channel.utils import create_channel
 from ..orders.utils.draft_order import draft_order_create
 from ..orders.utils.order_lines import order_lines_create
-from ..product.utils import (
-    create_category,
-    create_product,
-    create_product_channel_listing,
-    create_product_type,
-    create_product_variant,
-    create_product_variant_channel_listing,
-)
-from ..shipping_zone.utils import (
-    create_shipping_method,
-    create_shipping_method_channel_listing,
-    create_shipping_zone,
-)
+from ..product.utils.preparing_product import prepare_product
+from ..shop.utils.preparing_shop import prepare_shop
 from ..utils import assign_permissions
-from ..warehouse.utils import create_warehouse
 from .utils import checkout_create_from_order
-
-
-def prepare_product(
-    e2e_staff_api_client,
-    permission_manage_products,
-    permission_manage_channels,
-    permission_manage_shipping,
-    permission_manage_product_types_and_attributes,
-    permission_manage_orders,
-    permission_manage_checkouts,
-    channel_slug,
-):
-    permissions = [
-        permission_manage_products,
-        permission_manage_channels,
-        permission_manage_shipping,
-        permission_manage_product_types_and_attributes,
-        permission_manage_orders,
-        permission_manage_checkouts,
-    ]
-    assign_permissions(e2e_staff_api_client, permissions)
-
-    warehouse_data = create_warehouse(e2e_staff_api_client)
-    warehouse_id = warehouse_data["id"]
-
-    warehouse_ids = [warehouse_id]
-    channel_data = create_channel(
-        e2e_staff_api_client, slug=channel_slug, warehouse_ids=warehouse_ids
-    )
-    channel_id = channel_data["id"]
-
-    channel_ids = [channel_id]
-    shipping_zone_data = create_shipping_zone(
-        e2e_staff_api_client,
-        warehouse_ids=warehouse_ids,
-        channel_ids=channel_ids,
-    )
-    shipping_zone_id = shipping_zone_data["id"]
-
-    shipping_method_data = create_shipping_method(
-        e2e_staff_api_client, shipping_zone_id
-    )
-    shipping_method_id = shipping_method_data["id"]
-
-    create_shipping_method_channel_listing(
-        e2e_staff_api_client, shipping_method_id, channel_id
-    )
-
-    product_type_data = create_product_type(
-        e2e_staff_api_client,
-    )
-    product_type_id = product_type_data["id"]
-
-    category_data = create_category(e2e_staff_api_client)
-    category_id = category_data["id"]
-
-    product_data = create_product(e2e_staff_api_client, product_type_id, category_id)
-    product_id = product_data["id"]
-
-    create_product_channel_listing(e2e_staff_api_client, product_id, channel_id)
-
-    stocks = [
-        {
-            "warehouse": warehouse_id,
-            "quantity": 5,
-        }
-    ]
-    product_variant_data = create_product_variant(
-        e2e_staff_api_client,
-        product_id,
-        stocks=stocks,
-    )
-    product_variant_id = product_variant_data["id"]
-
-    create_product_variant_channel_listing(
-        e2e_staff_api_client,
-        product_variant_id,
-        channel_id,
-    )
-    data = draft_order_create(
-        e2e_staff_api_client,
-        channel_id,
-    )
-
-    order_id = data["order"]["id"]
-    order_lines = [{"variantId": product_variant_id, "quantity": 1, "price": 100}]
-    order_data = order_lines_create(e2e_staff_api_client, order_id, order_lines)
-    order_product_variant_id = order_data["order"]["lines"][0]["variant"]
-    order_product_quantity = order_data["order"]["lines"][0]["quantity"]
-
-    return (
-        order_id,
-        order_product_variant_id,
-        order_product_quantity,
-    )
 
 
 @pytest.mark.e2e
@@ -127,22 +19,39 @@ def test_checkout_create_from_order_core_0104(
     permission_manage_checkouts,
 ):
     # Before
-    (
-        order_id,
-        order_product_variant_id,
-        order_product_quantity,
-    ) = prepare_product(
-        e2e_staff_api_client,
+    permissions = [
         permission_manage_products,
         permission_manage_channels,
-        permission_manage_product_types_and_attributes,
         permission_manage_shipping,
+        permission_manage_product_types_and_attributes,
         permission_manage_orders,
         permission_manage_checkouts,
-        channel_slug="test",
+    ]
+    assign_permissions(e2e_staff_api_client, permissions)
+
+    (
+        result_warehouse_id,
+        result_channel_id,
+        _,
+        _,
+    ) = prepare_shop(e2e_staff_api_client)
+
+    product_variant_id = prepare_product(
+        e2e_staff_api_client, result_warehouse_id, result_channel_id
     )
 
     # Step 1 - Create checkout from order
+    data = draft_order_create(
+        e2e_staff_api_client,
+        result_channel_id,
+    )
+
+    order_id = data["order"]["id"]
+    assert order_id is not None
+    order_lines = [{"variantId": product_variant_id, "quantity": 1, "price": 100}]
+    order_data = order_lines_create(e2e_staff_api_client, order_id, order_lines)
+    order_product_variant_id = order_data["order"]["lines"][0]["variant"]
+    order_product_quantity = order_data["order"]["lines"][0]["quantity"]
 
     checkout_data = checkout_create_from_order(e2e_staff_api_client, order_id)
     checkout_id = checkout_data["checkout"]["id"]
