+++ saleor/tests/e2e/checkout/test_logged_customer_is_unable_to_buy_more_products_than_the_limit_per_customer.py
@@ -1,6 +1,5 @@
 import pytest
 
-from ..channel.utils import create_channel
 from ..product.utils import (
     create_category,
     create_product,
@@ -9,23 +8,17 @@ from ..product.utils import (
     create_product_variant,
     create_product_variant_channel_listing,
 )
-from ..shipping_zone.utils import (
-    create_shipping_method,
-    create_shipping_method_channel_listing,
-    create_shipping_zone,
-)
+from ..shop.utils.preparing_shop import prepare_shop
 from ..utils import assign_permissions
-from ..warehouse.utils import create_warehouse
 from .utils import checkout_create, checkout_lines_update
 
 
-def prepare_product(
+def prepare_product_with_limit(
     e2e_staff_api_client,
     permission_manage_products,
     permission_manage_channels,
     permission_manage_shipping,
     permission_manage_product_types_and_attributes,
-    channel_slug,
 ):
     permissions = [
         permission_manage_products,
@@ -35,30 +28,8 @@ def prepare_product(
     ]
     assign_permissions(e2e_staff_api_client, permissions)
 
-    warehouse_data = create_warehouse(e2e_staff_api_client)
-    warehouse_id = warehouse_data["id"]
-
-    warehouse_ids = [warehouse_id]
-    channel_data = create_channel(
-        e2e_staff_api_client, slug=channel_slug, warehouse_ids=warehouse_ids
-    )
-    channel_id = channel_data["id"]
-
-    channel_ids = [channel_id]
-    shipping_zone_data = create_shipping_zone(
-        e2e_staff_api_client,
-        warehouse_ids=warehouse_ids,
-        channel_ids=channel_ids,
-    )
-    shipping_zone_id = shipping_zone_data["id"]
-
-    shipping_method_data = create_shipping_method(
-        e2e_staff_api_client, shipping_zone_id
-    )
-    shipping_method_id = shipping_method_data["id"]
-
-    create_shipping_method_channel_listing(
-        e2e_staff_api_client, shipping_method_id, channel_id
+    result_warehouse_id, result_channel_id, result_channel_slug, _ = prepare_shop(
+        e2e_staff_api_client
     )
 
     product_type_data = create_product_type(
@@ -72,11 +43,11 @@ def prepare_product(
     product_data = create_product(e2e_staff_api_client, product_type_id, category_id)
     product_id = product_data["id"]
 
-    create_product_channel_listing(e2e_staff_api_client, product_id, channel_id)
+    create_product_channel_listing(e2e_staff_api_client, product_id, result_channel_id)
 
     stocks = [
         {
-            "warehouse": warehouse_id,
+            "warehouse": result_warehouse_id,
             "quantity": 5,
         }
     ]
@@ -95,13 +66,14 @@ def prepare_product(
     create_product_variant_channel_listing(
         e2e_staff_api_client,
         product_variant_id,
-        channel_id,
+        result_channel_id,
     )
 
     return (
         product_variant_id,
         product_variant_name,
         product_variant_quantity_limit_per_customer,
+        result_channel_slug,
     )
 
 
@@ -115,19 +87,17 @@ def test_checkout_with_product_quantity_exceeding_the_limit_per_customer_core_01
     permission_manage_product_types_and_attributes,
 ):
     # Before
-    channel_slug = "test-channel"
-
     (
         product_variant_id,
         product_variant_name,
         product_variant_quantity_limit_per_customer,
-    ) = prepare_product(
+        result_channel_slug,
+    ) = prepare_product_with_limit(
         e2e_staff_api_client,
         permission_manage_products,
         permission_manage_channels,
         permission_manage_shipping,
         permission_manage_product_types_and_attributes,
-        channel_slug,
     )
 
     # Step 1 - Create checkout.
@@ -137,7 +107,7 @@ def test_checkout_with_product_quantity_exceeding_the_limit_per_customer_core_01
     checkout_data = checkout_create(
         e2e_logged_api_client,
         lines,
-        channel_slug,
+        result_channel_slug,
         email="testEmail@example.com",
         set_default_billing_address=True,
     )
