+++ saleor/tests/e2e/checkout/test_logged_in_customer_should_not_be_able_to_buy_unavailable_product.py
@@ -1,6 +1,5 @@
 import pytest
 
-from ..channel.utils import create_channel
 from ..product.utils import (
     create_category,
     create_product,
@@ -9,29 +8,29 @@ from ..product.utils import (
     create_product_variant_channel_listing,
     raw_create_product_channel_listing,
 )
+from ..shop.utils.preparing_shop import prepare_shop
 from ..utils import assign_permissions
-from ..warehouse.utils import create_warehouse
 from .utils import raw_checkout_create
 
 
-def prepare_product(
+def prepare_unavailable_product(
     e2e_staff_api_client,
     permission_manage_products,
     permission_manage_channels,
     permission_manage_product_types_and_attributes,
+    permission_manage_shipping,
 ):
     permissions = [
         permission_manage_products,
         permission_manage_channels,
         permission_manage_product_types_and_attributes,
+        permission_manage_shipping,
     ]
 
     assign_permissions(e2e_staff_api_client, permissions)
-    warehouse_data = create_warehouse(e2e_staff_api_client)
-
-    channel_data = create_channel(e2e_staff_api_client, warehouse_data["id"])
-    channel_id = channel_data["id"]
-    channel_slug = channel_data["slug"]
+    result_warehouse_id, result_channel_id, result_channel_slug, _ = prepare_shop(
+        e2e_staff_api_client
+    )
 
     product_type_data = create_product_type(
         e2e_staff_api_client,
@@ -53,7 +52,7 @@ def prepare_product(
     product_channel_listing = raw_create_product_channel_listing(
         e2e_staff_api_client,
         product_id,
-        channel_id,
+        result_channel_id,
         is_available_for_purchase=False,
         available_for_purchase_datetime=None,
     )
@@ -64,7 +63,7 @@ def prepare_product(
 
     stocks = [
         {
-            "warehouse": warehouse_data["id"],
+            "warehouse": result_warehouse_id,
             "quantity": 5,
         }
     ]
@@ -76,10 +75,10 @@ def prepare_product(
     create_product_variant_channel_listing(
         e2e_staff_api_client,
         variant_id,
-        channel_id,
+        result_channel_id,
     )
 
-    return variant_id, channel_slug
+    return variant_id, result_channel_slug
 
 
 @pytest.mark.e2e
@@ -89,13 +88,15 @@ def test_should_not_be_able_to_buy_unavailable_product_core_0108(
     permission_manage_products,
     permission_manage_channels,
     permission_manage_product_types_and_attributes,
+    permission_manage_shipping,
 ):
     # Before
-    variant_id, channel_slug = prepare_product(
+    variant_id, channel_slug = prepare_unavailable_product(
         e2e_staff_api_client,
         permission_manage_products,
         permission_manage_channels,
         permission_manage_product_types_and_attributes,
+        permission_manage_shipping,
     )
     # Step 1 - Create checkout with non available for purchase product
     lines = [
