+++ saleor/tests/e2e/checkout/test_unlogged_customer_should_be_able_to_buy_by_click_and_collect.py
@@ -1,14 +1,7 @@
 import pytest
 
 from ..channel.utils import create_channel
-from ..product.utils import (
-    create_category,
-    create_product,
-    create_product_channel_listing,
-    create_product_type,
-    create_product_variant,
-    create_product_variant_channel_listing,
-)
+from ..product.utils.preparing_product import prepare_product
 from ..utils import assign_permissions
 from ..warehouse.utils import create_warehouse, update_warehouse
 from .utils import (
@@ -19,7 +12,7 @@ from .utils import (
 )
 
 
-def prepare_product(
+def prepare_shop_click_and_collect(
     e2e_staff_api_client,
     permission_manage_products,
     permission_manage_channels,
@@ -44,47 +37,11 @@ def prepare_product(
     channel_id = channel_data["id"]
     channel_slug = channel_data["slug"]
 
-    product_type_data = create_product_type(
-        e2e_staff_api_client,
-    )
-    product_type_id = product_type_data["id"]
-
-    category_data = create_category(
-        e2e_staff_api_client,
-    )
-    category_id = category_data["id"]
-
-    product_data = create_product(
-        e2e_staff_api_client,
-        product_type_id,
-        category_id,
-    )
-    product_id = product_data["id"]
-
-    create_product_channel_listing(e2e_staff_api_client, product_id, channel_id)
-
-    stocks = [
-        {
-            "warehouse": warehouse_data["id"],
-            "quantity": 5,
-        }
-    ]
-    variant_data = create_product_variant(
-        e2e_staff_api_client, product_id, stocks=stocks
-    )
-    variant_id = variant_data["id"]
-
-    create_product_variant_channel_listing(
-        e2e_staff_api_client,
-        variant_id,
-        channel_id,
-    )
-
-    return variant_id, channel_slug, warehouse_id
+    return channel_id, channel_slug, warehouse_id
 
 
 @pytest.mark.e2e
-def test_unlogged_customer_buy_by_click_and_collect(
+def test_unlogged_customer_buy_by_click_and_collect_CORE_0105(
     e2e_not_logged_api_client,
     e2e_staff_api_client,
     permission_manage_products,
@@ -92,16 +49,17 @@ def test_unlogged_customer_buy_by_click_and_collect(
     permission_manage_product_types_and_attributes,
 ):
     # Before
-    variant_id, channel_slug, warehouse_id = prepare_product(
+    channel_id, channel_slug, warehouse_id = prepare_shop_click_and_collect(
         e2e_staff_api_client,
         permission_manage_products,
         permission_manage_channels,
         permission_manage_product_types_and_attributes,
     )
+    product_variant_id = prepare_product(e2e_staff_api_client, warehouse_id, channel_id)
 
     # Step 1 - Create checkout and check collection point
     lines = [
-        {"variantId": variant_id, "quantity": 1},
+        {"variantId": product_variant_id, "quantity": 1},
     ]
     checkout_data = checkout_create(
         e2e_not_logged_api_client,
