+++ saleor/tests/e2e/checkout/test_unlogged_customer_should_be_able_to_order_digital_product.py
@@ -1,6 +1,5 @@
 import pytest
 
-from ..channel.utils import create_channel
 from ..product.utils import (
     create_category,
     create_digital_content,
@@ -10,9 +9,8 @@ from ..product.utils import (
     create_product_variant,
     create_product_variant_channel_listing,
 )
-from ..shipping_zone.utils import create_shipping_zone
+from ..shop.utils.preparing_shop import prepare_shop
 from ..utils import assign_permissions
-from ..warehouse.utils import create_warehouse
 from .utils import (
     checkout_billing_address_update,
     checkout_complete,
@@ -27,30 +25,16 @@ def prepare_product(
     permission_manage_channels,
     permission_manage_products,
     permission_manage_shipping,
-    channel_slug,
 ):
     permissions = [
-        permission_manage_products,
-        permission_manage_channels,
-        permission_manage_shipping,
         permission_manage_product_types_and_attributes,
+        permission_manage_channels,
+        permission_manage_products,
+        permission_manage_shipping,
     ]
     assign_permissions(e2e_staff_api_client, permissions)
-
-    warehouse_data = create_warehouse(e2e_staff_api_client)
-    warehouse_id = warehouse_data["id"]
-
-    warehouse_ids = [warehouse_id]
-    channel_data = create_channel(
-        e2e_staff_api_client, slug=channel_slug, warehouse_ids=warehouse_ids
-    )
-    channel_id = channel_data["id"]
-
-    channel_ids = [channel_id]
-    create_shipping_zone(
-        e2e_staff_api_client,
-        warehouse_ids=warehouse_ids,
-        channel_ids=channel_ids,
+    result_warehouse_id, result_channel_id, result_channel_slug, _ = prepare_shop(
+        e2e_staff_api_client
     )
 
     product_type_data = create_product_type(
@@ -66,11 +50,11 @@ def prepare_product(
     product_data = create_product(e2e_staff_api_client, product_type_id, category_id)
     product_id = product_data["id"]
 
-    create_product_channel_listing(e2e_staff_api_client, product_id, channel_id)
+    create_product_channel_listing(e2e_staff_api_client, product_id, result_channel_id)
 
     stocks = [
         {
-            "warehouse": warehouse_id,
+            "warehouse": result_warehouse_id,
             "quantity": 5,
         }
     ]
@@ -84,32 +68,29 @@ def prepare_product(
     create_product_variant_channel_listing(
         e2e_staff_api_client,
         product_variant_id,
-        channel_id,
+        result_channel_id,
     )
 
     create_digital_content(e2e_staff_api_client, product_variant_id)
-    return product_variant_id
+    return product_variant_id, result_channel_slug
 
 
 @pytest.mark.e2e
-def test_process_checkout_with_digital_product(
+def test_process_checkout_with_digital_product_CORE_0101(
     e2e_not_logged_api_client,
     e2e_staff_api_client,
     permission_manage_product_types_and_attributes,
     permission_manage_channels,
     permission_manage_products,
     permission_manage_shipping,
-    media_root,
 ):
     # Before
-    channel_slug = "test-channel"
-    product_variant_id = prepare_product(
+    product_variant_id, result_channel_slug = prepare_product(
         e2e_staff_api_client,
-        permission_manage_products,
-        permission_manage_channels,
-        permission_manage_shipping,
         permission_manage_product_types_and_attributes,
-        channel_slug,
+        permission_manage_channels,
+        permission_manage_products,
+        permission_manage_shipping,
     )
 
     # Step 1  - Create checkout.
@@ -119,7 +100,7 @@ def test_process_checkout_with_digital_product(
     checkout_data = checkout_create(
         e2e_not_logged_api_client,
         lines,
-        channel_slug,
+        result_channel_slug,
         email="testEmail@example.com",
     )
     checkout_id = checkout_data["id"]
