+++ saleor/tests/e2e/checkout/test_unlogged_customer_should_be_able_to_order_physical_product.py
@@ -1,21 +1,8 @@
 import pytest
 
-from ..channel.utils import create_channel
-from ..product.utils import (
-    create_category,
-    create_product,
-    create_product_channel_listing,
-    create_product_type,
-    create_product_variant,
-    create_product_variant_channel_listing,
-)
-from ..shipping_zone.utils import (
-    create_shipping_method,
-    create_shipping_method_channel_listing,
-    create_shipping_zone,
-)
+from ..product.utils.preparing_product import prepare_product
+from ..shop.utils.preparing_shop import prepare_shop
 from ..utils import assign_permissions
-from ..warehouse.utils import create_warehouse
 from .utils import (
     checkout_complete,
     checkout_create,
@@ -25,101 +12,39 @@ from .utils import (
 )
 
 
-def prepare_product(
+@pytest.mark.e2e
+def test_process_checkout_with_physical_product_CORE_0102(
     e2e_staff_api_client,
+    e2e_not_logged_api_client,
     permission_manage_products,
     permission_manage_channels,
-    permission_manage_shipping,
     permission_manage_product_types_and_attributes,
-    channel_slug,
+    permission_manage_shipping,
+    permission_manage_orders,
+    permission_manage_checkouts,
 ):
+    # Before
     permissions = [
         permission_manage_products,
         permission_manage_channels,
         permission_manage_shipping,
         permission_manage_product_types_and_attributes,
+        permission_manage_orders,
+        permission_manage_checkouts,
     ]
     assign_permissions(e2e_staff_api_client, permissions)
 
-    warehouse_data = create_warehouse(e2e_staff_api_client)
-    warehouse_id = warehouse_data["id"]
-
-    warehouse_ids = [warehouse_id]
-    channel_data = create_channel(
-        e2e_staff_api_client, slug=channel_slug, warehouse_ids=warehouse_ids
-    )
-    channel_id = channel_data["id"]
-
-    channel_ids = [channel_id]
-    shipping_zone_data = create_shipping_zone(
-        e2e_staff_api_client,
-        warehouse_ids=warehouse_ids,
-        channel_ids=channel_ids,
-    )
-    shipping_zone_id = shipping_zone_data["id"]
-
-    shipping_method_data = create_shipping_method(
-        e2e_staff_api_client, shipping_zone_id
-    )
-    shipping_method_id = shipping_method_data["id"]
-
-    create_shipping_method_channel_listing(
-        e2e_staff_api_client, shipping_method_id, channel_id
-    )
-
-    product_type_data = create_product_type(
-        e2e_staff_api_client,
-    )
-    product_type_id = product_type_data["id"]
-
-    category_data = create_category(e2e_staff_api_client)
-    category_id = category_data["id"]
-
-    product_data = create_product(e2e_staff_api_client, product_type_id, category_id)
-    product_id = product_data["id"]
-
-    create_product_channel_listing(e2e_staff_api_client, product_id, channel_id)
-
-    stocks = [
-        {
-            "warehouse": warehouse_id,
-            "quantity": 5,
-        }
-    ]
-    product_variant_data = create_product_variant(
-        e2e_staff_api_client,
-        product_id,
-        stocks=stocks,
-    )
-    product_variant_id = product_variant_data["id"]
-
-    create_product_variant_channel_listing(
-        e2e_staff_api_client,
-        product_variant_id,
+    (
+        warehouse_id,
         channel_id,
-    )
+        channel_slug,
+        shipping_method_id,
+    ) = prepare_shop(e2e_staff_api_client)
 
-    return product_variant_id
-
-
-@pytest.mark.e2e
-def test_process_checkout_with_physical_product(
-    e2e_staff_api_client,
-    e2e_not_logged_api_client,
-    permission_manage_products,
-    permission_manage_channels,
-    permission_manage_shipping,
-    permission_manage_product_types_and_attributes,
-):
-    # Before
-    channel_slug = "test-channel"
     product_variant_id = prepare_product(
         e2e_staff_api_client,
-        permission_manage_products,
-        permission_manage_channels,
-        permission_manage_shipping,
-        permission_manage_product_types_and_attributes,
-        channel_slug,
+        warehouse_id,
+        channel_id,
     )
 
     # Step 1 - Create checkout.
