+++ saleor/tests/e2e/checkout/test_unlogged_customer_should_not_be_able_to_order_product_quantity_greater_than_stock.py
@@ -1,6 +1,5 @@
 import pytest
 
-from ..channel.utils import create_channel
 from ..product.utils import (
     create_category,
     create_product,
@@ -9,9 +8,8 @@ from ..product.utils import (
     create_product_variant_channel_listing,
     raw_create_product_variant,
 )
-from ..shipping_zone.utils import create_shipping_zone
+from ..shop.utils.preparing_shop import prepare_shop
 from ..utils import assign_permissions
-from ..warehouse.utils import create_warehouse
 from .utils import raw_checkout_create
 
 
@@ -28,20 +26,14 @@ def prepare_product(
         permission_manage_product_types_and_attributes,
         permission_manage_shipping,
     ]
-
     assign_permissions(e2e_staff_api_client, permissions)
-    warehouse_data = create_warehouse(e2e_staff_api_client)
-    warehouse_id = warehouse_data["id"]
 
-    channel_data = create_channel(e2e_staff_api_client, warehouse_data["id"])
-    channel_id = channel_data["id"]
-    channel_slug = channel_data["slug"]
-
-    create_shipping_zone(
-        e2e_staff_api_client,
-        warehouse_ids=[warehouse_id],
-        channel_ids=[channel_id],
-    )
+    (
+        result_warehouse_id,
+        result_channel_id,
+        result_channel_slug,
+        _,
+    ) = prepare_shop(e2e_staff_api_client)
 
     product_type_data = create_product_type(
         e2e_staff_api_client,
@@ -61,26 +53,29 @@ def prepare_product(
     )
     product_id = product_data["id"]
 
-    create_product_channel_listing(e2e_staff_api_client, product_id, channel_id)
+    create_product_channel_listing(e2e_staff_api_client, product_id, result_channel_id)
+
+    stock_quantity = 5
 
     stocks = [
         {
-            "warehouse": warehouse_data["id"],
-            "quantity": 5,
+            "warehouse": result_warehouse_id,
+            "quantity": stock_quantity,
         }
     ]
     variant_data = raw_create_product_variant(
         e2e_staff_api_client, product_id, stocks=stocks
     )
-    variant_id = variant_data["productVariant"]["id"]
+    product_variant_id = variant_data["productVariant"]["id"]
+    product_variant_name = variant_data["productVariant"]["name"]
 
     create_product_variant_channel_listing(
         e2e_staff_api_client,
-        variant_id,
-        channel_id,
+        product_variant_id,
+        result_channel_id,
     )
 
-    return variant_id, channel_slug
+    return product_variant_id, product_variant_name, result_channel_slug, stock_quantity
 
 
 @pytest.mark.e2e
@@ -93,7 +88,12 @@ def test_unlogged_customer_cannot_buy_product_in_quantity_grater_than_stock_core
     permission_manage_shipping,
 ):
     # Before
-    variant_id, channel_slug = prepare_product(
+    (
+        product_variant_id,
+        product_variant_name,
+        result_channel_slug,
+        stock_quantity,
+    ) = prepare_product(
         e2e_staff_api_client,
         permission_manage_products,
         permission_manage_channels,
@@ -102,12 +102,12 @@ def test_unlogged_customer_cannot_buy_product_in_quantity_grater_than_stock_core
     )
     # Step 1 - Create checkout with product quantity greater than the available stock
     lines = [
-        {"variantId": variant_id, "quantity": 10},
+        {"variantId": product_variant_id, "quantity": stock_quantity + 1},
     ]
     checkout_data = raw_checkout_create(
         e2e_not_logged_api_client,
         lines,
-        channel_slug,
+        result_channel_slug,
         email="jon.doe@saleor.io",
         set_default_billing_address=True,
         set_default_shipping_address=True,
@@ -117,6 +117,6 @@ def test_unlogged_customer_cannot_buy_product_in_quantity_grater_than_stock_core
     assert errors[0]["code"] == "INSUFFICIENT_STOCK"
     assert errors[0]["field"] == "quantity"
     assert (
-        errors[0]["message"]
-        == "Could not add items Test product variant. Only 5 remaining in stock."
+        errors[0]["message"] == f"Could not add items {product_variant_name}. "
+        f"Only {stock_quantity} remaining in stock."
     )
