+++ saleor/tests/e2e/shop/utils/__init__.py
@@ -1,0 +1,7 @@
+from .preparing_shop import prepare_shop
+from .shop_update_settings import update_shop_settings
+
+__all__ = [
+    "prepare_shop",
+    "update_shop_settings",
+]
